// src/app/views/system/models/end-transactions.model.ts
export interface EndTransaction {
  cheque_amount?: number;
  cheque_quantity?: number;
  comment?: string;
  fifty_dollar_amount?: number;
  fifty_dollar_quantity?: number;
  five_cent_amount?: number;
  five_cent_quantity?: number;
  five_dollar_amount?: number;
  five_dollar_quantity?: number;
  id: number;
  id_balance: number;
  link_pay_amount?: number;
  link_pay_quantity?: number;
  location_id: number;
  one_cent_amount?: number;
  one_cent_quantity?: number;
  one_dollar_amount?: number;
  one_dollar_quantity?: number;
  one_hundred_dollar_amount?: number;
  one_hundred_dollar_quantity?: number;
  pos_amount?: number;
  pos_quantity?: number;
  start_balance: number;
  start_date: string;
  ten_cent_amount?: number;
  ten_cent_quantity?: number;
  ten_dollar_amount?: number;
  ten_dollar_quantity?: number;
  total_balance_detail?: number;
  transfer_bank_amount?: number;
  transfer_bank_quantity?: number;
  twenty_dollar_amount?: number;
  twenty_dollar_quantity?: number;
  twenty_five_cent_amount?: number;
  twenty_five_cent_quantity?: number;
  two_dollar_amount?: number;
  two_dollar_quantity?: number;
  user_id: string;
}